lambdas/requirements.txt
boto3==1.26.66
botocore==1.29.66
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
jmespath==1.0.1
multipart==0.2.4
Pillow==9.4.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
requests==2.28.2
requests-toolbelt==0.10.1
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.14
